Usage Instructions [WIP] [Feedback requested]
Hey guys, 

I think it would be good if we could add explanations of how the application works as we build out new features, since things aren't necessarily going to be obvious to us, and this would be good for turning in so a TA could actually use our tool.

Let me know what you think! Feel free to push to this branch with more details.